Refactor a project to different modules add MyException.java
