+++ Banking_Functionallity/src/main/java/dev/adrylat/bankingfunctionality/card/validators/CardValidator.java
@@ -1,0 +1,8 @@
+package dev.adrylat.bankingfunctionality.card.validators;
+
+import dev.adrylat.bankingfunctionality.card.MyException;
+
+public interface CardValidator {
+
+   boolean validate(String customerInput) throws MyException;
+}
