+++ Banking_Functionallity/src/main/java/dev/adrylat/bankingfunctionality/mortgagecalculator/MortgagePayment.java
@@ -1,0 +1,11 @@
+package dev.adrylat.bankingfunctionality.mortgagecalculator;
+
+public interface MortgagePayment {
+
+    void setMortgageDetails(CustomerData customerData);
+
+    void paymentCalculation();
+
+    double getPayment();
+
+}
