+++ Dialogs/src/main/java/dev/andrylat/dialogs/Dialog.java
@@ -1,0 +1,6 @@
+package dev.andrylat.dialogs;
+
+public interface Dialog {
+
+  void start();
+}
