--- src/main/java/dev/andrylat/bankingutilities/card/validators/CardValidator.java
@@ -1,10 +1,0 @@
-package dev.andrylat.bankingutilities.card.validators;
-
-import java.util.List;
-
-public interface CardValidator {
-
-    List<String> validate(String customerInput);
-
-
-}
