--- src/main/java/dev/andrylat/bankingutilities/dialogs/Dialog.java
@@ -1,6 +1,0 @@
-package dev.andrylat.bankingutilities.dialogs;
-
-public interface Dialog {
-
-  void start();
-}
