--- src/main/java/dev/andrylat/bankingutilities/mortgagecalculator/MortgagePayment.java
@@ -1,11 +1,0 @@
-package dev.andrylat.bankingutilities.mortgagecalculator;
-
-public interface MortgagePayment {
-
-    void setMortgageDetails(CustomerData customerData);
-
-    void paymentCalculation();
-
-    double getPayment();
-
-}
